add core movie module with CRUD endpoints and database seed scripts

+++ Project/src/main/java/com/furkan/project/auth/service/user/impl/UserService.java
@@ -143,7 +143,7 @@ public class UserService implements IUserService {
 
         return new PagedDataResult<>(
                 responses,
-                resultPage.getTotalElements(),
+                (int)resultPage.getTotalElements(),
                 resultPage.getTotalPages(),
                 resultPage.getNumber(),
                 resultPage.getSize(),
